requirements.txt
playwright==1.40.0
streamlit>=1.30.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
plotly>=5.18.0
pypdf>=3.17.0
